# optimize Inner::shallow_clone