- Clones when the kind is INLINE or STATIC are sped up by over double.
- Clones when the kind is ARC are spec up by about 10%. (Previously was 33%, but was using `ptr::copy` and TSAN warnings were triggered. While we *think* that would be safe because it's down only **after** an `Acquire` load, this PR opts to be conservative until we can verify it that it *certainly* is safe.)

Before

```
test clone_arc    ... bench:         27,175 ns/iter (+/- 1,195)
test clone_inline ... bench:         13,778 ns/iter (+/- 1,335)
test clone_static ... bench:         13,736 ns/iter (+/- 599)
```

After:

```
test clone_arc    ... bench:         25,398 ns/iter (+/- 1,209)
test clone_inline ... bench:          5,568 ns/iter (+/- 285)
test clone_static ... bench:          5,562 ns/iter (+/- 151)
```

## Warning

Modifies unsafe code, I could have goofed on something...