@seanmonstar 2018-06-26
Seems this triggers some TSAN warnings, since this is now calling `self.kind()` instead of `self.is_inline()` (which has a suppression configured). What's the preferred way to fix this? I can add a suppression to `kind`, but that might mask other unsafe uses of `kind`. I could add a `fn shallow_clone_kind` that is suppressed instead... Either way, I believe a race is not actually unsafe here, since the `kind` check is only trusted in safe cases (when inline or static), and otherwise, an atomic load is used.